Moved getSegmentBufferedPercent to `Ranges` module
