+++ test/segment-loader.test.js
@@ -2,43 +2,8 @@ import QUnit from 'qunit';
 import SegmentLoader from '../src/segment-loader';
 import videojs from 'video.js';
 import xhrFactory from '../src/xhr';
-import {useFakeEnvironment, useFakeMediaSource} from './test-helpers.js';
 import Config from '../src/config';
-
-const playlistWithDuration = function(time, conf) {
-  let result = {
-    targetDuration: 10,
-    mediaSequence: conf && conf.mediaSequence ? conf.mediaSequence : 0,
-    discontinuityStarts: [],
-    segments: [],
-    endList: true
-  };
-  let count = Math.floor(time / 10);
-  let remainder = time % 10;
-  let i;
-  let isEncrypted = conf && conf.isEncrypted;
-
-  for (i = 0; i < count; i++) {
-    result.segments.push({
-      uri: i + '.ts',
-      resolvedUri: i + '.ts',
-      duration: 10
-    });
-    if (isEncrypted) {
-      result.segments[i].key = {
-        uri: i + '-key.php',
-        resolvedUri: i + '-key.php'
-      };
-    }
-  }
-  if (remainder) {
-    result.segments.push({
-      uri: i + '.ts',
-      duration: remainder
-    });
-  }
-  return result;
-};
+import { playlistWithDuration, useFakeEnvironment, useFakeMediaSource } from './test-helpers.js';
 
 let currentTime;
 let mediaSource;
@@ -1100,150 +1065,4 @@ QUnit.test('doesn\'t allow more than one monitor buffer timer to be set', functi
   loader.monitorBuffer_();
 
   QUnit.equal(this.clock.methods.length, timeoutCount, 'timeout count remains the same');
-});
-
-QUnit.module('Segment Percent Buffered Calculations', {
-  beforeEach() {
-    this.env = useFakeEnvironment();
-    this.mse = useFakeMediaSource();
-    this.hasPlayed = true;
-    this.clock = this.env.clock;
-
-    currentTime = 0;
-    loader = new SegmentLoader({
-      currentTime() {
-        return currentTime;
-      },
-      mediaSource: new videojs.MediaSource(),
-      hasPlayed: () => this.hasPlayed
-    });
-  },
-  afterEach() {
-    this.env.restore();
-    this.mse.restore();
-  }
-});
-
-QUnit.test('calculates the percent buffered for segments', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  let buffered = videojs.createTimeRanges([[15, 19]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 0, buffered);
-
-  QUnit.equal(percentBuffered, 40, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered for segments taking into account ' +
-           'currentTime', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  let buffered = videojs.createTimeRanges([[15, 19]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 15, buffered);
-
-  QUnit.equal(percentBuffered, 90, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered for segments with multiple buffered ' +
-           'regions', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  let buffered = videojs.createTimeRanges([[0, 11], [12, 19]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 0, buffered);
-
-  QUnit.equal(percentBuffered, 80, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered for segments with multiple buffered ' +
-           'regions taking into account currentTime', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  let buffered = videojs.createTimeRanges([[0, 11], [12, 19]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 12, buffered);
-
-  QUnit.equal(percentBuffered, 90, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered for segments based on segment.end', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  playlist.segments[1].end = 19;
-
-  let buffered = videojs.createTimeRanges([[0, 19]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 0, buffered);
-
-  QUnit.equal(percentBuffered, 100, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered for segments based on segment.end taking ' +
-           'into account currentTime', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  playlist.segments[1].end = 19;
-
-  let buffered = videojs.createTimeRanges([[11, 20]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 11, buffered);
-
-  QUnit.equal(percentBuffered, 100, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered for segments based on segment.end with ' +
-           'multiple buffered regions', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  playlist.segments[1].end = 19;
-
-  let buffered = videojs.createTimeRanges([[0, 10], [12, 20]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 0, buffered);
-
-  QUnit.equal(percentBuffered, 80, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered for segments based on segment.end with ' +
-           'multiple buffered regions taking into account currentTime', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  playlist.segments[1].end = 19;
-
-  let buffered = videojs.createTimeRanges([[0, 10], [12, 20]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 12, buffered);
-
-  QUnit.equal(percentBuffered, 100, 'calculated the buffered amount correctly');
-});
-
-QUnit.test('calculates the percent buffered as 0 for zero-length segments', function() {
-  let playlist = playlistWithDuration(30);
-
-  loader.mimeType(this.mimeType);
-  loader.playlist(playlist);
-
-  playlist.segments[1].duration = 0;
-
-  let buffered = videojs.createTimeRanges([[0, 19]]);
-  let percentBuffered = loader.getSegmentBufferedPercent_(playlist, 1, 0, buffered);
-
-  QUnit.equal(percentBuffered, 0, 'calculated the buffered amount correctly');
 });
